requirements.txt

pyyaml
selenium>=4.0.0,<5.0.0
webdriver-manager>=4.0.0
pytest>=7.0.0
allure-pytest>=2.13.0